Cleanup: now in the IR we never store the AST nodes for classes
We store a new `IRClass` structure which contains the index and whether
the class is a statement or an expression. Similarly, in `ScriptNode`,
we don't need to store the `ClassDefNode` anymore, but we can simply
store the class count, which we'll use to build the index. Thus, the
backend (`CodeGenerator`) only needs to deal with the IR nodes.