Refactor: simplify stack traces code
We have currently various methods in the Evaluator interface used to
handle stack traces. They seem to be used only by RhinoException and I
believe the reason they are interface methods is just so that people can
ship a slimmed down version of rhino without the Interpreter, i.e.
shipping just the runtime and some classes compiled ahead-of-time.

If that's a use case we still want to support in 2.0, then I'll close
this PR, but if we decide that adding a few classes isn't such a big
deal anymore in 2026, we might use this as a starting point and clean up
more - there's a few other simplifications that I think are worth doing.

This came out internally as part of the new interpreter that we have
developed, and that we are working on upstreaming.